refactor: extract tab events from App
